+++ src/components/talk.js
@@ -1,4 +1,5 @@
 'use strict';
+import getTextStyle from "../textStyle.js";
 
 export default function talk(deliverer, options = { width: 100, offset: {x: 0, y: -20} }) {
 
@@ -26,17 +27,7 @@ export default function talk(deliverer, options = { width: 100, offset: {x: 0, y
                 text("", {
                     size: 6,
                     width: options.width - padding * 2,
-                    styles: {
-                        black: { color: rgb(0, 0, 0) },
-                        red: { color: rgb(255, 80, 77) },
-                        green: { color: rgb(66, 163, 121) },
-                        blue: { color: rgb(99, 155, 255) },
-                        yellow: { color: rgb(245, 182, 76) },
-                        pink: { color: rgb(229, 139, 199) },
-                        white: { color: rgb(203, 202, 174) },
-                        brown: { color: rgb(141, 82, 67) },
-                        grey: { color: rgb(92, 98, 120) }
-                    }
+                    styles: getTextStyle()
                 }),
                 origin("center"),
                 pos(box.pos.x, box.pos.y),