+++ src/textStyle.js
@@ -1,0 +1,15 @@
+
+
+const getTextStyle = () => ({
+    black: { color: rgb(0, 0, 0) },
+    red: { color: rgb(255, 80, 77) },
+    green: { color: rgb(66, 163, 121) },
+    blue: { color: rgb(99, 155, 255) },
+    yellow: { color: rgb(245, 182, 76) },
+    pink: { color: rgb(229, 139, 199) },
+    white: { color: rgb(203, 202, 174) },
+    brown: { color: rgb(141, 82, 67) },
+    grey: { color: rgb(92, 98, 120) }
+})
+
+export default getTextStyle