+++ CHANGELOG.md
@@ -3,10 +3,28 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.1.0 - 2022-25-10
+### Added
+- Added screen shaking when you try to give an order to the wrong client
+- Added mute shortcut with "M"
+- Added in-game tutorial and shortcut with "H"
+- Added restaurants names in map sprite
+- Added indications at the beginning of the game and at the first order.
+
+### Changed
+- Changed hint text "It's the GUY with ..." to be non-gendered. It can be assumed that we are referring to men only
+- Changed score/points to money "$" to be more realistic.
+- Associate one food to each restaurant (Big Donald = Burger, KCF = Chicken, Chuchichop = Sushi, Chef Michel = meal, Gromino's = Pizza)
+- Increased default music volume
+- Changed food sprites to be more harmonious
+- Changed margin between picked up food orders
+- Rotate more often NPC to see their colors better
+- Changed the color of the restaurant name in the order ready message
+- Changed space bar icon to be more visible with white outline
+
 ## v1.0.1 - 2022-10-25
 ### Fixed
 - Increase game over text Z-index
 
 ## v1.0.0 - 2022-10-05
-### Added
-First version, nothing to say here.
+Official version published for Ludum Dare 51.